requirements.txt
requests
mysql-connector
selenium